+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 Version 20.1-SNAPSHOT
 -------------
 
+ADDED:
+
+- new method in SetFactory with collection to exclude
+  
 Version 20
 -------------